# Improve the speed of parallel safe random streams